chore: refactor mui theme to use factories
