--- src/theme/index.ts
@@ -1,1 +1,0 @@
-export { darkTheme, lightTheme } from './theme';
