+++ src/theme/shape.ts
@@ -1,0 +1,13 @@
+import type { Shape } from '@mui/material/styles';
+
+export const defaultShape: Shape = {
+  borderRadius: 12,
+  borderRadiusSecondary: 8,
+  borderRadiusTertiary: 4,
+  tabBarRadius: 128,
+  cardBorderRadius: 24,
+  buttonBorderRadius: 128,
+  inputTextBorderRadius: 24,
+};
+
+export const shape = defaultShape;
