+++ src/theme/typography.ts
@@ -1,0 +1,311 @@
+import type { Breakpoint, Theme } from '@mui/material/styles';
+import { inter, urbanist } from 'src/fonts/fonts';
+
+export interface ThemeFonts {
+  body: string;
+  display: string;
+}
+
+export const defaultFonts: ThemeFonts = {
+  body: inter.style.fontFamily,
+  display: urbanist.style.fontFamily,
+};
+
+export const createTypography = (
+  themeBase: Theme,
+  fonts: ThemeFonts = defaultFonts,
+) => ({
+  fontFamily: [
+    fonts.body,
+    'Arial',
+    'Noto Sans',
+    'BlinkMacSystemFont',
+    'Segoe UI',
+    'Helvetica Neue',
+    'sans-serif',
+  ].join(','),
+  headerXLarge: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '64px',
+    lineHeight: '96px',
+    letterSpacing: 0,
+  },
+  headerLarge: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '48px',
+    lineHeight: '64px',
+    letterSpacing: 0,
+  },
+  headerMedium: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '32px',
+    lineHeight: '40px',
+    letterSpacing: 0,
+  },
+  headerSmall: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '18px',
+    lineHeight: '24px',
+    letterSpacing: 0,
+  },
+  headerXSmall: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '16px',
+    lineHeight: '20px',
+    letterSpacing: 0,
+  },
+  bodyXLargeStrong: {
+    fontStyle: 'normal',
+    fontWeight: 800,
+    fontSize: '24px',
+    lineHeight: '32px',
+    letterSpacing: 0,
+  },
+  bodyXLarge: {
+    fontStyle: 'normal',
+    fontWeight: 400,
+    fontSize: '24px',
+    lineHeight: '32px',
+    letterSpacing: 0,
+  },
+  bodyLargeStrong: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '18px',
+    lineHeight: '24px',
+    letterSpacing: 0,
+  },
+  bodyLarge: {
+    fontStyle: 'normal',
+    fontWeight: 500,
+    fontSize: '18px',
+    lineHeight: '24px',
+    letterSpacing: 0,
+  },
+  bodyMediumStrong: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '16px',
+    lineHeight: '20px',
+    letterSpacing: 0,
+  },
+  bodyMedium: {
+    fontStyle: 'normal',
+    fontWeight: 500,
+    fontSize: '16px',
+    lineHeight: '20px',
+    letterSpacing: 0,
+  },
+  bodyMediumParagraph: {
+    fontStyle: 'normal',
+    fontWeight: 500,
+    fontSize: '16px',
+    lineHeight: '24px',
+    letterSpacing: 0,
+  },
+  bodySmallStrong: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '14px',
+    lineHeight: '18px',
+    letterSpacing: 0,
+  },
+  bodySmallParagraph: {
+    fontStyle: 'normal',
+    fontWeight: 500,
+    fontSize: '14px',
+    lineHeight: '20px',
+    letterSpacing: 0,
+  },
+  bodySmall: {
+    fontStyle: 'normal',
+    fontWeight: 400,
+    fontSize: '14px',
+    lineHeight: '20px',
+    letterSpacing: 0,
+  },
+  bodyXSmallStrong: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '12px',
+    lineHeight: '16px',
+    letterSpacing: 0,
+  },
+  bodyXSmall: {
+    fontStyle: 'normal',
+    fontWeight: 500,
+    fontSize: '12px',
+    lineHeight: '16px',
+    letterSpacing: 0,
+  },
+  bodyXXSmallStrong: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '10px',
+    lineHeight: '14px',
+    letterSpacing: 0,
+  },
+  bodyXXSmall: {
+    fontStyle: 'normal',
+    fontWeight: 400,
+    fontSize: '10px',
+    lineHeight: '14px',
+    letterSpacing: 0,
+  },
+  brandHeaderXLarge: {
+    fontStyle: 'normal',
+    fontWeight: 700,
+    fontSize: '64px',
+    lineHeight: '72px',
+    letterSpacing: 0,
+  },
+  titleSmall: {
+    fontFamily: fonts.body,
+    fontSize: '24px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '32px',
+  },
+  title2XSmall: {
+    fontFamily: fonts.body,
+    fontSize: '14px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '20px',
+  },
+  titleMedium: {
+    fontFamily: fonts.body,
+    fontSize: '32px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '40px',
+  },
+  titleLarge: {
+    fontSize: '48px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '64px',
+  },
+  title2XLarge: {
+    fontFamily: fonts.body,
+    fontSize: '64px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '72px',
+  },
+  titleXSmall: {
+    fontFamily: fonts.body,
+    fontSize: '18px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '24px',
+  },
+  urbanistTitleXSmall: {
+    fontFamily: fonts.display,
+    fontSize: '18px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '24px',
+  },
+  urbanistTitleLarge: {
+    fontFamily: fonts.display,
+    fontSize: '48px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '56px',
+  },
+  urbanistTitleXLarge: {
+    fontFamily: fonts.display,
+    fontSize: '64px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '72px',
+  },
+  urbanistTitle2XLarge: {
+    fontFamily: fonts.display,
+    fontSize: '80px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '96px',
+  },
+  urbanistTitle3XLarge: {
+    fontSize: '96px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '112px',
+  },
+  urbanistTitleMedium: {
+    fontFamily: fonts.display,
+    fontSize: '32px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '40px',
+  },
+  urbanistBodyLarge: {
+    fontFamily: fonts.display,
+    fontSize: '18px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '24px',
+  },
+  urbanistBodyXLarge: {
+    fontFamily: fonts.display,
+    fontSize: '24px',
+    fontStyle: 'normal',
+    fontWeight: 400,
+    lineHeight: '32px',
+  },
+  urbanistBody2XLarge: {
+    fontFamily: fonts.display,
+    fontSize: '32px',
+    fontStyle: 'normal',
+    fontWeight: 700,
+    lineHeight: '40px',
+  },
+  h1: {
+    fontFamily: fonts.display,
+    fontSize: themeBase.typography.pxToRem(48),
+    lineHeight: themeBase.typography.pxToRem(64),
+    letterSpacing: 'inherit',
+    fontWeight: 700,
+    [themeBase.breakpoints.up('sm' as Breakpoint)]: {
+      fontSize: themeBase.typography.pxToRem(64),
+      lineHeight: themeBase.typography.pxToRem(72),
+    },
+  },
+  h2: {
+    fontFamily: fonts.display,
+    fontSize: themeBase.typography.pxToRem(36),
+    lineHeight: themeBase.typography.pxToRem(48),
+    fontWeight: 700,
+  },
+  h3: {
+    fontFamily: fonts.display,
+    fontSize: themeBase.typography.pxToRem(28),
+    lineHeight: themeBase.typography.pxToRem(36),
+    fontWeight: 700,
+  },
+  h4: {
+    fontFamily: fonts.display,
+    fontSize: themeBase.typography.pxToRem(22),
+    lineHeight: themeBase.typography.pxToRem(28),
+    fontWeight: 700,
+  },
+  h5: {
+    fontFamily: fonts.display,
+    fontSize: themeBase.typography.pxToRem(18),
+    lineHeight: themeBase.typography.pxToRem(24),
+    fontWeight: 700,
+  },
+  h6: {
+    fontFamily: fonts.display,
+    fontSize: themeBase.typography.pxToRem(12),
+    lineHeight: themeBase.typography.pxToRem(18),
+    fontWeight: 700,
+  },
+});
